# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: FedProspect API
  version: "1.0.0"
  description: >
    Federal contract prospecting platform API. Provides endpoints for searching
    federal opportunities, awards, entities, managing prospects, proposals,
    saved searches, organization profiles, and administrative functions.
servers:
  - url: https://localhost:5001/api
    description: Local development server
...
security:
  - cookieAuth: []
paths:
...
  /v1/opportunities:
    get:
      tags: [Opportunities]
      summary: Search opportunities
      description: Search federal opportunities with filters and pagination.
      operationId: searchOpportunities
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: naics
          in: query
          schema:
            type: string
        - name: setAside
          in: query
          schema:
            type: string
        - name: solicitation
          in: query
          schema:
            type: string
        - name: daysOut
          in: query
          schema:
            type: integer
        - name: openOnly
          in: query
          schema:
            type: boolean
            default: true
        - name: department
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/sortBy"
        - $ref: "#/components/parameters/sortDescending"
      responses:
        "200":
          description: Paged list of opportunities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedOpportunitySearchResponse"
        "401":
          description: Not authenticated
  /v1/opportunities/targets:
    get:
      tags: [Opportunities]
      summary: Get target opportunities
      description: Search target opportunities filtered by value range, NAICS sector, and other criteria.
      operationId: getTargetOpportunities
      parameters:
        - name: minValue
          in: query
          schema:
            type: number
            format: decimal
        - name: maxValue
          in: query
          schema:
            type: number
            format: decimal
        - name: naicsSector
          in: query
          schema:
            type: string
        - name: setAside
          in: query
          schema:
            type: string
        - name: naics
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/sortBy"
        - $ref: "#/components/parameters/sortDescending"
      responses:
        "200":
          description: Paged list of target opportunities
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Not authenticated
  /v1/opportunities/export:
    get:
      tags: [Opportunities]
      summary: Export opportunities as CSV
      description: Export opportunity search results as a CSV file. Accepts the same parameters as the search endpoint.
      operationId: exportOpportunitiesCsv
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: naics
          in: query
          schema:
            type: string
        - name: setAside
          in: query
          schema:
            type: string
        - name: solicitation
          in: query
          schema:
            type: string
        - name: daysOut
          in: query
          schema:
            type: integer
        - name: openOnly
          in: query
          schema:
            type: boolean
            default: true
        - name: department
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/sortBy"
        - $ref: "#/components/parameters/sortDescending"
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary
        "401":
          description: Not authenticated
  /v1/opportunities/recommended:
    get:
      tags: [Opportunities]
      summary: Get recommended opportunities
      description: Get opportunities scored and ranked for the current organization's profile.
      operationId: getRecommendedOpportunities
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of recommended opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecommendedOpportunityDto"
        "401":
          description: Not authenticated
  /v1/opportunities/pwin/batch:
    post:
      tags: [Opportunities]
      summary: Calculate batch pWin
      description: Calculate probability of win for multiple opportunities in one request (max 200).
      operationId: calculateBatchPWin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchPWinRequest"
      responses:
        "200":
          description: Batch pWin results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchPWinResponse"
        "400":
          description: Too many notice IDs (max 200)
        "401":
          description: Not authenticated
  /v1/opportunities/{noticeId}:
    get:
      tags: [Opportunities]
      summary: Get opportunity detail
      description: Get full detail for a single opportunity by notice ID.
      operationId: getOpportunityDetail
      parameters:
        - name: noticeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Opportunity detail
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Not authenticated
        "404":
          description: Opportunity not found
  /v1/opportunities/{noticeId}/pwin:
    get:
      tags: [Opportunities]
      summary: Calculate pWin
      description: Calculate probability of win for an opportunity against the current organization's profile.
      operationId: calculatePWin
      parameters:
        - name: noticeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: pWin result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PWinResultDto"
        "401":
          description: Not authenticated
  /v1/opportunities/{noticeId}/incumbent:
    get:
      tags: [Opportunities]
      summary: Get incumbent analysis
      description: Identifies current contract holder and vulnerability signals for an opportunity.
      operationId: getIncumbentAnalysis
      parameters:
        - name: noticeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Incumbent analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncumbentAnalysisDto"
        "401":
          description: Not authenticated
  /v1/opportunities/{noticeId}/competitive-landscape:
    get:
      tags: [Opportunities]
      summary: Get competitive landscape
      description: Get agency + NAICS scoped market data with fallback for an opportunity.
      operationId: getCompetitiveLandscape
      parameters:
        - name: noticeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Competitive landscape data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitiveLandscapeDto"
        "401":
          description: Not authenticated
        "404":
          description: No competitive data available
  /v1/opportunities/{noticeId}/set-aside-shift:
    get:
      tags: [Opportunities]
      summary: Get set-aside shift analysis
      description: Compares current set-aside to predecessor contract for an opportunity.
      operationId: getSetAsideShift
      parameters:
        - name: noticeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Set-aside shift data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetAsideShiftDto"
        "401":
          description: Not authenticated
        "404":
          description: No predecessor data available
  /v1/opportunities/{noticeId}/qualification:
    get:
      tags: [Opportunities]
      summary: Run qualification checks
      description: Run qualification checks for an opportunity against the current organization's profile.
      operationId: checkQualification
      parameters:
        - name: noticeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Qualification check results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualificationCheckDto"
        "401":
          description: Not authenticated
  /v1/opportunities/{noticeId}/document-intelligence:
    get:
      tags: [Opportunities]
      summary: Get document intelligence
      description: Get intelligence extracted from opportunity attachments (SOW, RFP, etc.).
      operationId: getDocumentIntelligence
      parameters:
        - name: noticeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document intelligence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentIntelligenceDto"
        "401":
          description: Not authenticated
        "404":
          description: No document intelligence available
  /v1/opportunities/{noticeId}/analyze:
    post:
      tags: [Opportunities]
      summary: Request attachment analysis
      description: Request AI analysis of opportunity attachments. Inserts a data_load_request for the Python pipeline.
      operationId: requestAnalysis
      parameters:
        - name: noticeId
          in: path
          required: true
          schema:
            type: string
        - name: tier
          in: query
          schema:
            type: string
            default: haiku
            enum: [haiku, sonnet]
      responses:
        "200":
          description: Load request status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoadRequestStatusDto"
        "401":
          description: Not authenticated
  /v1/opportunities/{noticeId}/fetch-description:
    post:
      tags: [Opportunities]
      summary: Fetch missing description from SAM.gov
      description: >
        Fetch a missing opportunity description from SAM.gov on demand. If already
        populated, returns the cached text without making an API call. Includes SSRF
        protection (blocks private IPs), HTML stripping, and DB caching.
      operationId: fetchDescription
      parameters:
        - name: noticeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Description text (cached or freshly fetched)
          content:
            application/json:
              schema:
                type: object
                properties:
                  noticeId:
                    type: string
                  descriptionText:
                    type: string
        "401":
          description: Not authenticated
        "404":
          description: Opportunity not found or not found on SAM.gov
        "502":
          description: Upstream SAM.gov call failed

  # ---------------------------------------------------------------------------
  # Awards
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: httpOnly JWT cookie set by the login endpoint
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 25
        minimum: 1
        maximum: 100
    sortBy:
      name: sortBy
      in: query
      schema:
        type: string
    sortDescending:
      name: sortDescending
      in: query
      schema:
        type: boolean
        default: false
  schemas:
...
    PagedOpportunitySearchResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer
        hasPreviousPage:
          type: boolean
        hasNextPage:
          type: boolean
...
    PWinResultDto:
      type: object
      properties:
        prospectId:
          type: integer
        noticeId:
          type: string
        score:
          type: number
          format: decimal
        category:
          type: string
        factors:
          type: array
          items:
            $ref: "#/components/schemas/PWinFactorDto"
        suggestions:
          type: array
          items:
            type: string
    PWinFactorDto:
      type: object
      properties:
        name:
          type: string
        score:
          type: number
          format: decimal
        weight:
          type: number
          format: decimal
        weightedScore:
          type: number
          format: decimal
        detail:
          type: string
    BatchPWinRequest:
      type: object
      required: [noticeIds]
      properties:
        noticeIds:
          type: array
          items:
            type: string
          maxItems: 200
    BatchPWinResponse:
      type: object
      properties:
        results:
          type: object
          additionalProperties:
            type: object
            properties:
              score:
                type: number
                format: decimal
              category:
                type: string
    RecommendedOpportunityDto:
      type: object
      properties:
        noticeId:
          type: string
        title:
          type: string
        solicitationNumber:
          type: string
        departmentName:
          type: string
        subTier:
          type: string
        setAsideCode:
          type: string
        setAsideDescription:
          type: string
        naicsCode:
          type: string
        naicsDescription:
          type: string
        classificationCode:
          type: string
        noticeType:
          type: string
        awardAmount:
          type: number
          format: decimal
        postedDate:
          type: string
          format: date-time
        responseDeadline:
          type: string
          format: date-time
        daysRemaining:
          type: integer
        popState:
          type: string
        popCity:
          type: string
        popCountry:
          type: string
        qScore:
          type: number
          format: decimal
        qScoreCategory:
          type: string
        qScoreFactors:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              points:
                type: number
                format: decimal
              maxPoints:
                type: number
                format: decimal
        isRecompete:
          type: boolean
        incumbentName:
          type: string
    IncumbentAnalysisDto:
      type: object
      properties:
        noticeId:
          type: string
        hasIncumbent:
          type: boolean
        incumbentUei:
          type: string
        incumbentName:
          type: string
        contractId:
          type: string
        contractValue:
          type: number
          format: decimal
        dollarsObligated:
          type: number
          format: decimal
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        monthsRemaining:
          type: integer
        monthlyBurnRate:
          type: number
          format: decimal
        percentSpent:
          type: number
          format: decimal
        registrationStatus:
          type: string
        registrationExpiration:
          type: string
          format: date-time
        isExcluded:
          type: boolean
        exclusionType:
          type: string
        totalContractsInNaics:
          type: integer
        consecutiveWins:
          type: integer
        vulnerabilitySignals:
          type: array
          items:
            type: string
        isLikelyIncumbent:
          type: boolean
        likelyCompetitors:
          type: array
          items:
            type: object
            properties:
              vendorName:
                type: string
              ueiSam:
                type: string
              contractCount:
                type: integer
              totalValue:
                type: number
                format: decimal
    CompetitiveLandscapeDto:
      type: object
      properties:
        naicsCode:
          type: string
        agencyCode:
          type: string
        setAsideCode:
          type: string
        totalContracts:
          type: integer
        totalValue:
          type: number
          format: decimal
        averageAwardValue:
          type: number
          format: decimal
        agencyAverageAwardValue:
          type: number
          format: decimal
        topVendors:
          type: array
          items:
            $ref: "#/components/schemas/VendorShareDto"
        competitionLevel:
          type: string
        distinctVendorCount:
          type: integer
        fallbackScope:
          type: string
...
    VendorShareDto:
      type: object
      properties:
        vendorUei:
          type: string
        vendorName:
          type: string
        contractCount:
          type: integer
        totalValue:
          type: number
          format: decimal
        marketSharePercent:
          type: number
          format: decimal
    SetAsideShiftDto:
      type: object
      properties:
        noticeId:
          type: string
        solicitationNumber:
          type: string
        currentSetAsideCode:
          type: string
        currentSetAsideDescription:
          type: string
        predecessorSetAsideType:
          type: string
        predecessorVendorName:
          type: string
        predecessorVendorUei:
          type: string
        predecessorDateSigned:
          type: string
          format: date-time
        predecessorValue:
          type: number
          format: decimal
        shiftDetected:
          type: boolean
...
    QualificationCheckDto:
      type: object
      properties:
        noticeId:
          type: string
        passCount:
          type: integer
        failCount:
          type: integer
        warningCount:
          type: integer
        totalChecks:
          type: integer
        overallStatus:
          type: string
          enum: [Qualified, Partially Qualified, Not Qualified]
        checks:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              category:
                type: string
                enum: [Certification, Experience, Compliance, Logistics]
              status:
                type: string
                enum: [Pass, Fail, Warning, Unknown]
              detail:
                type: string
              sourceUei:
                type: string
    DocumentIntelligenceDto:
      type: object
      properties:
        noticeId:
          type: string
        attachmentCount:
          type: integer
        analyzedCount:
          type: integer
        latestExtractionMethod:
          type: string
        lastExtractedAt:
          type: string
          format: date-time
        clearanceRequired:
          type: string
        clearanceLevel:
          type: string
        clearanceScope:
          type: string
        evalMethod:
          type: string
        vehicleType:
          type: string
        isRecompete:
          type: string
        incumbentName:
          type: string
        scopeSummary:
          type: string
        periodOfPerformance:
          type: string
        laborCategories:
          type: array
          items:
            type: string
        keyRequirements:
          type: array
          items:
            type: string
        overallConfidence:
          type: string
          enum: [low, medium, high]
        sources:
          type: array
          items:
            type: object
            properties:
              fieldName:
                type: string
              sourceFilename:
                type: string
              pageNumber:
                type: integer
              matchedText:
                type: string
              surroundingContext:
                type: string
              extractionMethod:
                type: string
              confidence:
                type: string
        attachments:
          type: array
          items:
            type: object
            properties:
              attachmentId:
                type: integer
              filename:
                type: string
              contentType:
                type: string
              fileSizeBytes:
                type: integer
                format: int64
              pageCount:
                type: integer
              downloadStatus:
                type: string
              extractionStatus:
                type: string
    LoadRequestStatusDto:
      type: object
      properties:
        requestId:
          type: integer
        requestType:
          type: string
        status:
          type: string
        requestedAt:
          type: string
          format: date-time
        errorMessage:
          type: string
...